move beans to data directory
