+++ src/components/Course/index.tsx
@@ -12,7 +12,7 @@ import {
 import { classes, getContentClassName } from '../../utils';
 import { ActionRow, Instructor, Palette, Prerequisite } from '..';
 import { TermContext } from '../../contexts';
-import { Course as CourseBean, Section } from '../../beans';
+import { Course as CourseBean, Section } from '../../data/beans';
 import { CourseGpa, CrawlerPrerequisites } from '../../types';
 import { ErrorWithFields, softError } from '../../log';
