+++ src/components/CourseAdd/index.tsx
@@ -14,7 +14,7 @@ import { Course, CourseFilter } from '..';
 import { classes, getRandomColor } from '../../utils';
 import { ASYNC_DELIVERY_MODE, CAMPUSES, DELIVERY_MODES } from '../../constants';
 import { TermContext } from '../../contexts';
-import { Course as CourseBean, Section } from '../../beans';
+import { Course as CourseBean, Section } from '../../data/beans';
 
 import './stylesheet.scss';
