+++ src/data/beans/Feedback.ts
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-import { BACKEND_BASE_URL } from '../constants';
+import { BACKEND_BASE_URL } from '../../constants';
 
 export default async function FormSubmit({
   rating,
