+++ src/utils.tsx
@@ -5,7 +5,7 @@ import { getSanitizedOptions } from 'exponential-backoff/dist/options';
 import domtoimage from 'dom-to-image';
 import { saveAs } from 'file-saver';
 
-import { Oscar, Section } from './beans';
+import { Oscar, Section } from './data/beans';
 import { DAYS, PALETTE, PNG_SCALE_FACTOR } from './constants';
 import { ErrorWithFields, softError } from './log';
 import { ICS, Period, PrerequisiteClause, Theme } from './types';
